Extract module execution order analysis from Graph
